Split target state set/get into separate module
This removes the circular dependency between device state index and
preloading and cleans up the device state module code

Change-type: patch
